requirements.txt
alembic==1.7.7
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Mako==1.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
six==1.16.0
SQLAlchemy==1.4.34
Werkzeug==2.1.1